fix bug with showing category with todos, add new 6 colors, started new page for category creation and updation

+++ app/src/main/java/com/easycodingstudio/todou/data/database/dao/TodoDao.kt
@@ -9,7 +9,7 @@ import kotlinx.coroutines.flow.Flow
 interface TodoDao {
 
     @Query("SELECT * FROM todos WHERE category_id = :categoryId")
-    fun getTodosForCategory(categoryId: Int): Flow<List<TodoEntity>>
+    fun getTodosForCategory(categoryId: Long): Flow<List<TodoEntity>>
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insert(entity: TodoEntity): Long
